changing orientation does not cause issues, standard atlas formatting of conversation items in conversationlistactivity applied
